legacy/pyproject.toml
[project]
name = "aptamers-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.12"
dependencies = [
    "absl-py==2.2.2",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "alembic==1.16.4",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "asttokens==3.0.0",
    "astunparse==1.6.3",
    "async-timeout==5.0.1",
    "attrs==25.3.0",
    "beautifulsoup4==4.13.5",
    "biopython==1.85",
    "blinker==1.9.0",
    "cachetools==5.5.2",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "click==8.2.1",
    "cloudpickle==3.1.1",
    "colorama==0.4.6",
    "colorlog==6.9.0",
    "comm==0.2.2",
    "contourpy==1.3.0",
    "cryptography==45.0.7",
    "cycler==0.12.1",
    "databricks-sdk==0.65.0",
    "debugpy==1.8.14",
    "decorator==5.2.1",
    "docker==7.1.0",
    "exceptiongroup==1.2.2",
    "executing==2.2.0",
    "fastapi==0.116.1",
    "filelock==3.18.0",
    "flask==3.1.2",
    "flatbuffers==25.2.10",
    "fonttools==4.57.0",
    "frozenlist==1.7.0",
    "fsspec==2025.3.2",
    "gast==0.6.0",
    "gitdb==4.0.12",
    "gitpython==3.1.45",
    "google-auth==2.40.3",
    "google-pasta==0.2.0",
    "graphene==3.4.3",
    "graphql-core==3.2.6",
    "graphql-relay==3.2.0",
    "graphviz==0.20.3",
    "greenlet==3.2.3",
    "grpcio==1.71.0",
    "gunicorn==23.0.0",
    "h11==0.16.0",
    "h5py==3.13.0",
    "huggingface-hub==0.30.2",
    "hydra-core>=1.3.2",
    "idna==3.10",
    "importlib-metadata==8.6.1",
    "importlib-resources==6.5.2",
    "iniconfig==2.1.0",
    "ipykernel==6.29.5",
    "ipython==8.18.1",
    "itsdangerous==2.2.0",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "joblib==1.5.2",
    "jupyter-client==8.6.3",
    "jupyter-core==5.7.2",
    "kagglehub==0.3.12",
    "keras==3.9.2",
    "keras-hub==0.20.0",
    "keras-nlp==0.20.0",
    "kiwisolver==1.4.7",
    "libclang==18.1.1",
    "mako==1.3.10",
    "markdown==3.8",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "matplotlib==3.9.4",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "ml-dtypes==0.5.1",
    "mlflow==3.3.2",
    "mlflow-skinny==3.3.2",
    "mlflow-tracing==3.3.2",
    "mpmath==1.3.0",
    "multidict==6.6.4",
    "namex==0.0.8",
    "nest-asyncio==1.6.0",
    "networkx==3.2.1",
    "numpy>=2.1.3",
    "opentelemetry-api==1.36.0",
    "opentelemetry-sdk==1.36.0",
    "opentelemetry-semantic-conventions==0.57b0",
    "opt-einsum==3.4.0",
    "optree==0.15.0",
    "optuna==4.4.0",
    "packaging==24.2",
    "pandas==2.2.3",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pillow==11.2.1",
    "platformdirs==4.3.7",
    "pluggy==1.6.0",
    "prompt-toolkit==3.0.51",
    "propcache==0.3.2",
    "protobuf==5.29.4",
    "psutil==7.0.0",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyarrow==21.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydot==3.0.4",
    "pygments==2.19.1",
    "pyparsing==3.2.3",
    "pytest==8.4.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv>=1.2.1",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "pyzmq==26.4.0",
    "regex==2024.11.6",
    "requests==2.32.3",
    "rich==14.0.0",
    "rsa==4.9.1",
    "safetensors==0.5.3",
    "scikit-learn==1.7.1",
    "scipy==1.15.3",
    "seaborn==0.13.2",
    "sentencepiece==0.2.0",
    "seqshannon==1.0.0",
    "six==1.17.0",
    "smmap==5.0.2",
    "snakeviz==2.2.2",
    "sniffio==1.3.1",
    "soupsieve==2.8",
    "sqlalchemy==2.0.42",
    "sqlparse==0.5.3",
    "stack-data==0.6.3",
    "starlette==0.47.3",
    "sympy==1.13.1",
    "tensorboard==2.19.0",
    "tensorboard-data-server==0.7.2",
    "tensorflow==2.19.0",
    "tensorflow-text==2.19.0",
    "termcolor==3.0.1",
    "threadpoolctl==3.6.0",
    "tokenizers==0.21.1",
    "tomli==2.2.1",
    "torch==2.6.0",
    "torchinfo==1.8.0",
    "torchsummary==1.5.1",
    "tornado==6.4.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "transformers==4.51.3",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "uvicorn==0.35.0",
    "wcwidth==0.2.13",
    "werkzeug==3.1.3",
    "wrapt==1.17.2",
    "xgboost==3.1.1",
    "yarl==1.20.1",
    "zipp==3.21.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
